+++ NEWS.md
@@ -1,4 +1,4 @@
-# nanonext 0.8.0.9001 (development)
+# nanonext 0.8.1
 
 #### New Features
 
@@ -6,13 +6,14 @@
   + adds core functions `cv()`, `wait()`, `until()`, `cv_value()`, and `cv_reset()`.
   + adds signalling receive functions `recv_aio_signal()` and `request_signal()`.
   + `pipe_notify()` signals up to 2 condition variables whenever pipes are added or removed at a socket.
-* Implements `msg_pipe()` to return the pipe connection associated with a 'recvAio' message.
+* Adds `msg_pipe()` to return the pipe connection associated with a 'recvAio' message.
 * Exposes the `sha1()` cryptographic hash and HMAC generation function from the 'Mbed TLS' library (for secure applications, use one of the SHA-2 algorithms instead).
+* Utility function `'weakref<-'()` exposes `R_MakeWeakRef` from R's C API. Useful for keeping objects alive for as long as required by a dependent object.
 
 #### Updates
 
+* `ncurl_session()` gains a 'timeout' argument, and returns an 'errorValue' with warning upon error.
 * `listen()` and `dial()` gain the new logical argument 'error' to govern the function behaviour upon error.
-* `.unresolved()` is introduced as a technical utility (not recommended for general use) that just queries the status of an Aio without attempting to retrieve the result or message.
 * Internal performance enhancements.
 
 # nanonext 0.8.0